Move config to dir (easier docker config editing)
