+++ .gitignore
@@ -1,9 +1,9 @@
 node_modules/
 
-/app/config.*
-!/app/config.example.js
-/server/config.*
-!/server/config.example.js
+/app/config/
+!/app/config/config.example.js
+/server/config/
+!/server/config/config.example.js
 /server/public/
 /server/certs/
 !/server/certs/mediasoup-demo.localhost.*
