+++ server/config/config.example.js
@@ -11,8 +11,8 @@ module.exports =
 	domain : 'localhost',
 	tls    :
 	{
-		cert : `${__dirname}/certs/mediasoup-demo.localhost.cert.pem`,
-		key  : `${__dirname}/certs/mediasoup-demo.localhost.key.pem`
+		cert : `${__dirname}/../certs/mediasoup-demo.localhost.cert.pem`,
+		key  : `${__dirname}/../certs/mediasoup-demo.localhost.key.pem`
 	},
 	// Listening port for https server.
 	listeningPort : 443,
